requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
ImageHash==4.3.1
loguru==0.7.3
cnocr==2.3.1
cnstd==1.2.5.2
numpy==1.26.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==11.1.0
PyYAML==6.0.2
uiautomator2==3.2.8